Refactoring part 2
 #1 Backend process-lifecycle helpers (composition, not inheritance)
  - New WebUI/electron/subprocesses/processLifecycle.ts exporting
    terminateProcessTree() (SIGTERM → Windows taskkill /T /F tree / SIGKILL)
    and waitForServerReadyOrThrow() (health poll; supports llamacpp's
    getStartupError and openvino's stderr/exit-signal diagnostics).
  - service.ts (base stop()), openVINOBackendService.ts and
    llamaCppBackendService.ts now delegate to these instead of each
    reimplementing health-check / kill logic.
  - Fix: llamacpp LLM/embedding stops previously used SIGTERM→SIGKILL with no
    Windows tree-kill, leaking descendant processes and causing reinstall
    EPERM; they now tree-kill like the other backends.

  #2 Shared Python loopback-auth
  - New backend_shared/aipg_loopback_auth.py: stdlib-only, transport-agnostic
    evaluate_loopback_auth() + get_loopback_token().
  - service/web_api.py and home-agent/web_api.py import it via a
    ../backend_shared sys.path bootstrap (resolves identically in dev and
    packaged layouts); each keeps its own response/CORS handling so behavior
    is preserved.
  - build-config.json: package backend_shared/ alongside the backends.

  #3 Split comfyUiPresets store (1600 → 1368 lines)
  - New comfyUiMessages.ts: ComfyUI WS message schema + execution-error
    summarizer.
  - New comfyUiWorkflowTransforms.ts: workflow-graph transforms (OVMS image
    url inject, model-path normalize, node bypass).
  - comfyUiPresets.ts imports both; reactive store surface unchanged.
