+++ WebUI/electron/subprocesses/openVINOBackendService.ts
@@ -13,6 +13,10 @@ import { ensureManagedPython, installBackend, uvPipInstallToTarget } from './uvB
 import { binary, extract } from './tools.ts'
 import { getBundledBackendVersionSync, resolveModels } from '../remoteUpdates.ts'
 import {
+  terminateProcessTree as killProcessTree,
+  waitForServerReadyOrThrow,
+} from './processLifecycle.ts'
+import {
   getMissingPackages,
   hasAptGet,
   hasPkexec,
@@ -2136,39 +2140,7 @@ export class OpenVINOBackendService implements ApiService {
    * whole tree; we then wait for the OS to reap it and release the handles.
    */
   private async terminateProcessTree(proc: ChildProcess, label: string): Promise<void> {
-    const waitForExit = (ms: number): Promise<boolean> =>
-      new Promise<boolean>((resolve) => {
-        if (proc.exitCode !== null || proc.signalCode !== null) {
-          resolve(true)
-          return
-        }
-        const timeout = setTimeout(() => resolve(false), ms)
-        proc.once('exit', () => {
-          clearTimeout(timeout)
-          resolve(true)
-        })
-      })
-
-    proc.kill('SIGTERM')
-    let exited = await waitForExit(2000)
-    if (exited) return
-
-    this.appLogger.warn(`Force killing OVMS ${label} process tree`, this.name)
-    if (process.platform === 'win32' && proc.pid !== undefined) {
-      try {
-        await execAsync(`taskkill /PID ${proc.pid} /T /F`)
-      } catch (e) {
-        // taskkill exits non-zero when the process is already gone — not fatal.
-        this.appLogger.warn(`taskkill for OVMS ${label} reported: ${e}`, this.name)
-      }
-    } else {
-      proc.kill('SIGKILL')
-    }
-
-    exited = await waitForExit(5000)
-    if (!exited) {
-      this.appLogger.warn(`OVMS ${label} not confirmed exited after force kill`, this.name)
-    }
+    await killProcessTree(proc, { name: this.name, label, appLogger: this.appLogger })
   }
 
   private async stopOvmsLlmServer(): Promise<void> {
@@ -2676,91 +2648,12 @@ export class OpenVINOBackendService implements ApiService {
     childProcess: ChildProcess,
     maxAttempts = 120,
   ): Promise<void> {
-    const delayMs = 1000
-
-    // Track whether the process has exited (process.killed only reflects
-    // signals sent by Node.js — not OS-level kills like OOM or SIGSEGV).
-    let processExited = false
-    let exitCode: number | null = null
-    let exitSignal: string | null = null
-    const stderrChunks: string[] = []
-
-    const onExit = (code: number | null, signal: string | null) => {
-      processExited = true
-      exitCode = code
-      exitSignal = signal
-    }
-    childProcess.on('exit', onExit)
-    childProcess.stderr?.on('data', (data: Buffer) => {
-      stderrChunks.push(data.toString())
-      // Keep only the last 20 lines of stderr for diagnostics
-      if (stderrChunks.length > 20) stderrChunks.shift()
+    await waitForServerReadyOrThrow(healthUrl, childProcess, {
+      name: this.name,
+      maxAttempts,
+      captureExitDiagnostics: true,
+      appLogger: this.appLogger,
     })
-
-    const buildExitErrorMessage = (): string => {
-      const reason = exitSignal
-        ? `killed by signal ${exitSignal}`
-        : exitCode !== null
-          ? `exit code ${exitCode}`
-          : 'exit code null (killed by OS signal, possibly OOM)'
-      const lastStderr = stderrChunks.join('').trim()
-      const stderrSuffix = lastStderr ? `\nLast stderr output:\n${lastStderr.slice(-2000)}` : ''
-      return `OVMS process crashed during startup (${reason})${stderrSuffix}`
-    }
-
-    try {
-      for (let attempt = 0; attempt < maxAttempts; attempt++) {
-        // Check if process has exited (covers OS kills, crashes, OOM, etc.)
-        if (processExited || childProcess.killed) {
-          const msg = buildExitErrorMessage()
-          this.appLogger.warn(
-            `Process for ${this.name} is not alive, aborting health check: ${msg}`,
-            this.name,
-          )
-          throw new Error(msg)
-        }
-
-        try {
-          const response = await fetch(healthUrl, {
-            method: 'GET',
-            signal: AbortSignal.timeout(1000),
-          })
-
-          if (response.ok) {
-            if (processExited || childProcess.killed) {
-              const msg = buildExitErrorMessage()
-              this.appLogger.warn(
-                `Process for ${this.name} exited after health check succeeded: ${msg}`,
-                this.name,
-              )
-              throw new Error(msg)
-            }
-            this.appLogger.info(`Server ready at ${healthUrl}`, this.name)
-            return
-          }
-        } catch (error) {
-          // Re-throw our own errors (from the process-exit check above)
-          if (error instanceof Error && error.message.startsWith('OVMS process crashed')) {
-            throw error
-          }
-          // Server not ready yet — check if the process died while we were waiting
-          if (processExited || childProcess.killed) {
-            const msg = buildExitErrorMessage()
-            this.appLogger.warn(
-              `Process for ${this.name} exited during health check wait: ${msg}`,
-              this.name,
-            )
-            throw new Error(msg)
-          }
-        }
-
-        await new Promise((resolve) => setTimeout(resolve, delayMs))
-      }
-
-      throw new Error(`Server failed to start within ${(maxAttempts * delayMs) / 1000} seconds`)
-    } finally {
-      childProcess.removeListener('exit', onExit)
-    }
   }
 
   // Error management methods for startup failures
