+++ WebUI/electron/subprocesses/service.ts
@@ -6,6 +6,7 @@ import path from 'node:path'
 import { appLoggerInstance } from '../logging/logger.ts'
 import { packagedResourcesRoot } from '../aipgRoot.ts'
 import { existingFileOrError, spawnProcessAsync, ProcessError } from './osProcessHelper'
+import { terminateProcessTree } from './processLifecycle.ts'
 import { assert } from 'node:console'
 import { createHash } from 'crypto'
 
@@ -716,48 +717,7 @@ export abstract class LongLivedPythonApiService implements ApiService {
       return 'stopped'
     }
 
-    const waitForExit = (ms: number): Promise<boolean> =>
-      Promise.race([
-        new Promise<boolean>((resolve) => proc.once('exit', () => resolve(true))),
-        new Promise<boolean>((resolve) => setTimeout(() => resolve(false), ms)),
-      ])
-
-    // Try graceful shutdown first with SIGTERM
-    proc.kill('SIGTERM')
-
-    // Wait up to 2 seconds for the process to exit gracefully
-    let exited = await waitForExit(2000)
-
-    if (!exited) {
-      this.appLogger.warn(
-        `Backend ${this.name} did not exit gracefully within 2s, force killing`,
-        this.name,
-      )
-      // On Windows, ChildProcess.kill() only signals the direct child and leaves
-      // descendant processes (e.g. ComfyUI's python and the uv subprocesses spawned
-      // by ComfyUI-Manager) running. Those descendants keep handles on the service
-      // directory — and python's CWD is the service dir itself — which makes a
-      // subsequent reinstall fail to delete it (EPERM). taskkill /T /F tears down
-      // the whole tree.
-      if (process.platform === 'win32' && proc.pid !== undefined) {
-        try {
-          await exec(`taskkill /PID ${proc.pid} /T /F`)
-        } catch (e) {
-          // taskkill exits non-zero when the process is already gone — not fatal.
-          this.appLogger.warn(`taskkill for backend ${this.name} reported: ${e}`, this.name)
-        }
-      } else {
-        proc.kill('SIGKILL')
-      }
-      // Wait for the OS to actually reap the process and release file handles.
-      exited = await waitForExit(5000)
-      if (!exited) {
-        this.appLogger.warn(
-          `Backend ${this.name} still not confirmed exited after force kill`,
-          this.name,
-        )
-      }
-    }
+    await terminateProcessTree(proc, { name: this.name, appLogger: this.appLogger })
 
     this.encapsulatedProcess = null
     this.setStatus('stopped')
