+++ WebUI/src/assets/js/store/comfyUiPresets.ts
@@ -15,7 +15,6 @@ import { useActivities } from './activities'
 import { createAppError } from '../errors/appError'
 import { useBackendServices } from '@/assets/js/store/backendServices.ts'
 import { usePromptStore } from './promptArea'
-import { z } from 'zod'
 import { imageUrlToDataUri, isImageUrl, mediaUrl } from '@/lib/utils'
 import { getComfyAuthToken, invalidateComfyAuthToken } from '@/lib/loopbackAuth'
 import {
@@ -23,6 +22,17 @@ import {
   findKeysByTitle,
   modifySettingInWorkflow,
 } from './comfyUiWorkflowHelpers'
+import {
+  ComfyMessageSchema,
+  summarizeComfyExecutionError,
+  type ComfyExecutionErrorData,
+} from './comfyUiMessages'
+import {
+  bypassNode,
+  injectOvmsImageUrl,
+  normalizeModelPathsInWorkflow,
+  workflowUsesOvmsImage,
+} from './comfyUiWorkflowTransforms'
 
 /**
  * Wraps fetch() with the ComfyUI loopback bearer token. The bundled
@@ -49,248 +59,6 @@ async function comfyFetch(input: RequestInfo | URL, init?: RequestInit): Promise
 }
 
 const WEBSOCKET_OPEN = 1
-
-const ComfyMessageSchema = z.discriminatedUnion('type', [
-  z.object({
-    type: z.literal('status'),
-  }),
-  z.object({
-    type: z.literal('execution_start'),
-    data: z.object({}).passthrough(),
-  }),
-  z.object({
-    type: z.literal('execution_success'),
-    data: z.object({}).passthrough(),
-  }),
-  z.object({
-    type: z.literal('execution_error'),
-    data: z
-      .object({
-        exception_message: z.string().optional(),
-        exception_type: z.string().optional(),
-        node_id: z.union([z.string(), z.number()]).optional(),
-        node_type: z.string().optional(),
-        traceback: z.array(z.string()).optional(),
-      })
-      .passthrough(),
-  }),
-  z.object({
-    type: z.literal('execution_interrupted'),
-    data: z.object({}).passthrough(),
-  }),
-  z.object({
-    type: z.literal('execution_cached'),
-    data: z.object({}).passthrough(),
-  }),
-  z.object({
-    type: z.literal('progress'),
-    data: z
-      .object({
-        value: z.number(),
-        max: z.number(),
-      })
-      .passthrough(),
-  }),
-  z.object({
-    type: z.literal('executing'),
-    data: z
-      .object({
-        node: z.string().nullable().optional(),
-        display_node: z.string().optional(),
-      })
-      .passthrough(),
-  }),
-  z.object({
-    type: z.literal('executed'),
-    data: z
-      .object({
-        output: z.union([
-          z.object({
-            images: z.array(
-              z.object({
-                filename: z.string(),
-                subfolder: z.string(),
-                type: z.string(),
-              }),
-            ),
-            animated: z.array(z.boolean()).optional(),
-          }),
-          z.object({
-            gifs: z.array(
-              z.object({
-                filename: z.string(),
-                workflow: z.string(),
-                type: z.string(),
-                subfolder: z.string(),
-                format: z.string(),
-              }),
-            ),
-          }),
-          z.object({
-            '3d': z.array(
-              z.object({
-                filename: z.string(),
-                subfolder: z.string(),
-                type: z.string(),
-              }),
-            ),
-          }),
-        ]),
-      })
-      .passthrough(),
-  }),
-  z.object({
-    type: z.literal('progress_state'),
-    data: z.object({}).passthrough(),
-  }),
-])
-
-type ComfyExecutionErrorData = {
-  exception_message?: string
-  exception_type?: string
-  node_id?: string | number
-  node_type?: string
-  traceback?: string[]
-}
-
-// ComfyUI reports node failures with a full Python exception (often a multi-KB
-// state_dict / size-mismatch dump). That is useless and overwhelming as a
-// user-facing string, so we map the common, recognizable failures to a short,
-// actionable sentence and keep the raw detail for the logs/debug panel only.
-function summarizeComfyExecutionError(data: ComfyExecutionErrorData): string {
-  const raw = (data.exception_message ?? '').trim()
-  const lower = raw.toLowerCase()
-  const type = (data.exception_type ?? '').toLowerCase()
-
-  if (
-    lower.includes('size mismatch') ||
-    lower.includes('error(s) in loading state_dict') ||
-    lower.includes('load_state_dict')
-  ) {
-    return "The selected model doesn't match this workflow (mismatched weights while loading). Pick a model that fits the preset, or choose a different preset."
-  }
-  if (
-    lower.includes('out of memory') ||
-    lower.includes('outofmemory') ||
-    lower.includes('failed to allocate') ||
-    (lower.includes('alloc') && lower.includes('memory'))
-  ) {
-    return 'Ran out of memory while generating. Try a smaller resolution or batch size.'
-  }
-  if (
-    type.includes('filenotfound') ||
-    lower.includes('no such file') ||
-    lower.includes('cannot find') ||
-    lower.includes('does not exist')
-  ) {
-    return 'A required model or file could not be found. Make sure the needed models are downloaded.'
-  }
-
-  // Fallback: first meaningful line of the exception, trimmed to a sane length.
-  const firstLine =
-    raw
-      .split('\n')
-      .map((line) => line.trim())
-      .find((line) => line.length > 0) ?? ''
-  const concise = firstLine.length > 200 ? `${firstLine.slice(0, 200)}…` : firstLine
-  const prefix = data.node_type ? `${data.node_type}: ` : ''
-  return concise ? `${prefix}${concise}` : 'The workflow failed during execution.'
-}
-
-const OVMS_IMAGE_CLASS_TYPES = ['OpenAICompatibleImageGeneration', 'OpenAICompatibleImageEdit']
-
-function workflowUsesOvmsImage(workflow: ComfyUIApiWorkflow): boolean {
-  return OVMS_IMAGE_CLASS_TYPES.some((ct) => findKeysByClassType(workflow, ct).length > 0)
-}
-
-function injectOvmsImageUrl(workflow: ComfyUIApiWorkflow, url: string): void {
-  for (const classType of OVMS_IMAGE_CLASS_TYPES) {
-    for (const key of findKeysByClassType(workflow, classType)) {
-      const inputs = workflow[key]?.inputs
-      if (!inputs || typeof inputs !== 'object') continue
-      if ('base_url' in inputs) {
-        inputs['base_url'] = url
-      }
-      // OVMS registers the served graph under the slash-flattened repo id
-      // (see `--source_model` in openVINOBackendService.ts), so the model
-      // sent in the OpenAI-compatible request must use the same form.
-      if ('model' in inputs && typeof inputs['model'] === 'string') {
-        inputs['model'] = inputs['model'].split('/').join('---')
-      }
-    }
-  }
-}
-
-/** ComfyUI node input names that hold model/file paths; separator is OS-dependent (see main.ts preset handling). */
-const COMFY_MODEL_PATH_INPUTS = new Set([
-  'ckpt_name',
-  'lora_name',
-  'text_encoder',
-  'vae_name',
-  'unet_name',
-  'clip_name',
-  'model_name',
-  'control_net_name',
-])
-
-function normalizeModelPathsInWorkflow(
-  workflow: ComfyUIApiWorkflow,
-  platform: NodeJS.Platform,
-): void {
-  for (const node of Object.values(workflow)) {
-    const inputs = (node as { inputs?: Record<string, unknown> }).inputs
-    if (!inputs) continue
-    for (const [inputName, value] of Object.entries(inputs)) {
-      if (COMFY_MODEL_PATH_INPUTS.has(inputName) && typeof value === 'string') {
-        inputs[inputName] = modelNameForComfyApi(value, platform)
-      }
-    }
-  }
-}
-
-/**
- * Bypass a node by rewiring its outputs to its upstream and removing the node.
- * Supported: LoraLoader (output 0 = model from input "model", output 1 = clip from input "clip").
- */
-function bypassNode(workflow: ComfyUIApiWorkflow, nodeId: string): void {
-  const node = workflow[nodeId] as
-    | { class_type?: string; inputs?: Record<string, unknown> }
-    | undefined
-  if (!node?.inputs) return
-  const classType = node.class_type
-  let rewire: [number, [string, number]][]
-  if (classType === 'LoraLoader') {
-    const model = node.inputs.model as [string, number] | undefined
-    const clip = node.inputs.clip as [string, number] | undefined
-    if (!model || !clip) return
-    rewire = [
-      [0, model],
-      [1, clip],
-    ]
-  } else {
-    return
-  }
-  for (const entry of Object.values(workflow)) {
-    const inputs = (entry as { inputs?: Record<string, unknown> }).inputs
-    if (!inputs) continue
-    for (const key of Object.keys(inputs)) {
-      const v = inputs[key]
-      if (
-        Array.isArray(v) &&
-        v.length === 2 &&
-        typeof v[0] === 'string' &&
-        typeof v[1] === 'number'
-      ) {
-        if (v[0] === nodeId) {
-          const slot = v[1]
-          const upstream = rewire.find(([s]) => s === slot)?.[1]
-          if (upstream) inputs[key] = upstream
-        }
-      }
-    }
-  }
-  delete workflow[nodeId]
-}
 
 /** Rewire and remove optional model nodes whose value is None (e.g. LoRA bypass). */
 function bypassOptionalModelNodes(workflow: ComfyUIApiWorkflow): void {
